moved classes for easier merge
